feat: `worktree::encode_to_worktree()` to turn UTf-8 into the worktree encoding.
